We now have 3 PR's related to sorting logits with the goal to speedup top-k sampling:
* PR #5085 - speed up via usage of `std::nth_element` before `std::partial_sort`
* PR #5101, which introduces a bucket sort algorithm
* This PR, which provides a better (faster) implementation of the bucket sort proposed in #5101 

The table shows a comparison between master and these 3 PR's as a function of `top_k`. Tests run on a Ryzen-5975WX + RTX--4080 with Ubuntu 22.04 and GCC 11.4.0.

| --top-k | t/s master | t/s (PR 5085) | t/s PR 5101 | t/s this PR | Speedup this PR |
|---------|------------|-----------------|---------------------|-----------------|---------------------|
|      100 |       7758 |            7758 |   3084 | 7758 | 1.000 |
|      200 |       6796 |            6796 |   3076 | 7121 |  1.048 |
|      500 |       4848 |            4848 |    2905 | 6733 | 1.389 |
|    1000 |       3353 |            3353 |    2872 | 6075 | 1.812 |
|    2000 |       2123 |            2123 |    2592 | 5006 | 2.358 |
|    4000 |       1288 |            2000 |    2246 | 3627 | 2.816 |
|    8000 |         783 |            1438 |    1718 | 2292 | 2.927 |
|   16000 |        508 |             921 |     1143 | 1326 | 2.611 |
|   31780 |        397 |             558 |       675 | 734  | 1.848 |
|   32000 |        559 |             559 |       703 | 773 |  1.384 |