@cmp-nct 2024-01-24
If this continues we'll soon see a GPU tensor-core sorting kernel beating this one again :)
It's quite amazing to see how much performance can be gained.
On the other hand, I think pre-filtering K to a more meaningful value is still the best way to go in terms of practicality, if huge-top-k values are the future.