@ikawrakow 2024-01-24
> If this continues we'll soon see a GPU tensor-core sorting kernel beating this one again :) It's quite amazing to see how much performance can be gained. On the other hand, I think pre-filtering K to a more meaningful value is still the best way to go in terms of practicality, if huge-top-k values are the future.

Well, when the GPU beats this, there is still room for improvement. One can easily shave off another 10% or so from the time by having a top_k sampler instance that has the buffers pre-allocated, so one doesn't need to do memory allocations on each invocation of `llama_sample_top_k`. One can vectorize at least the 1st loop. Etc.

More seriously, I do agree with you that if the usage if large top_k becomes standard practice, it is better to prefilter the logits in some way.