@JohannesGaessler 2024-01-24
PR for a min_p implementation that works on unsorted tokens: https://github.com/ggerganov/llama.cpp/pull/5115 .